chore: delete unused code
