--- app/client/src/ce/sagas/EnvironmentSagas.ts
@@ -1,3 +1,0 @@
-export function* waitForFetchEnvironments() {
-  yield true;
-}
