--- app/client/src/ce/utils/analyticsHelpers.ts
@@ -1,7 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-export const logMainJsActionExecution = (
-  actionId: string,
-  isSuccess: boolean,
-  collectionId: string,
-  isDirty: boolean,
-) => {};
