--- app/client/src/ce/utils/autocomplete/helpers.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import type { Def } from "tern";
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const eeAppsmithAutocompleteDefs = (generatedTypeDef: Def) => ({});
