+++ app/client/src/ee/utils/analyticsHelpers.ts
@@ -1,1 +1,7 @@
-export * from "ce/utils/analyticsHelpers";
+/* eslint-disable @typescript-eslint/no-unused-vars */
+export const logMainJsActionExecution = (
+  actionId: string,
+  isSuccess: boolean,
+  collectionId: string,
+  isDirty: boolean,
+) => {};
