+++ app/client/src/ee/utils/autocomplete/helpers.ts
@@ -1,1 +1,5 @@
-export * from "ce/utils/autocomplete/helpers";
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import type { Def } from "tern";
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const eeAppsmithAutocompleteDefs = (generatedTypeDef: Def) => ({});
